NODE_RESERVED_NAMES_CONST {data.tree}	R Documentation

Names that are reserved by the Node class.

Description

These are reserved by the Node class, you cannot use these as attribute names. Note also that all fields starting with a . are reserved.

Usage

NODE_RESERVED_NAMES_CONST

Format

An object of class character of length 40.

[Package data.tree version 0.7.5 Index]