as.Node.dendrogram {data.tree}	R Documentation

Convert a dendrogram to a data.tree Node

Description

Convert a dendrogram to a data.tree Node

Usage

## S3 method for class 'dendrogram'
as.Node(x, name = "Root", heightName = "plotHeight",
  check = c("check", "no-warn", "no-check"), ...)

Arguments

x
The dendrogram

name
The name of the root Node

heightName
The name under which the dendrogram's height is stored

check
Either

"check": if the name conformance should be checked and warnings should be printed in case of non-conformance (the default)

"no-warn": if the name conformance should be checked, but no warnings should be printed in case of non-conformance (if you expect non-conformance)

"no-check" or FALSE: if the name conformance should not be checked; use this if performance is critical. However, in case of non-conformance, expect cryptic follow-up errors

...
Additional parameters

Value

The root Node of a data.tree

See Also

Other as.Node: as.Node.data.frame, as.Node.list, as.Node.phylo, as.Node

Examples

hc <- hclust(dist(USArrests), "ave")
dend1 <- as.dendrogram(hc)
tree1 <- as.Node(dend1)
tree1$fieldsAll
tree1$totalCount
tree1$leafCount
tree1$height
  

[Package data.tree version 0.7.5 Index]