as.Node.phylo {data.tree}	R Documentation

Convert a phylo object from the ape package to a Node

Description

Convert a phylo object from the ape package to a Node

Usage

## S3 method for class 'phylo'
as.Node(x, heightName = "plotHeight",
  replaceUnderscores = TRUE, ...)

Arguments

x
The phylo object to be converted

heightName
If the phylo contains edge lengths, then they will be converted to a height and stored in a field named according to this parameter (the default is "height")

replaceUnderscores
if TRUE (the default), then underscores in names are replaced with spaces

...
any other parameter to be passed to sub-implementations

See Also

Other ape phylo conversions: GetPhyloNr, as.phylo.Node

Other as.Node: as.Node.data.frame, as.Node.dendrogram, as.Node.list, as.Node

Examples

#which bird familes have the max height?
library(ape)
data(bird.families)
bf <- as.Node(bird.families)
height <- bf$height
t <- Traverse(bf, filterFun = function(x) x$level == 25)
Get(t, "name")

[Package data.tree version 0.7.4 Index]