as.igraph.Node {data.tree}	R Documentation

Convert a data.tree structure to an igraph network

Description

This requires the igraph package to be installed. Also, this requires the names of the Nodes to be unique within the data.tree structure.

Usage

as.igraph.Node(node, vertexAttributes = character(),
  edgeAttributes = character(), directed = FALSE,
  direction = c("climb", "descend"))

Arguments

node
The root Node to convert

vertexAttributes
A vector of strings, representing the attributes in the data.tree structure to add as attributes to the vertices of the igraph

edgeAttributes
A vector of strings, representing the attributes in the data.tree structure to add as edge attributes of the igraph

directed
Logical scalar, whether or not to create a directed graph.

direction
when converting to a network, should the edges point from root to children ("climb") or from child to parent ("descend")?

Value

an igraph object

See Also

AreNamesUnique

Examples

data(acme)
library(igraph)
ig <- as.igraph(acme, "p", c("level", "isLeaf"))
plot(ig)

[Package data.tree version 0.7.7 Index]