as.list.Node {data.tree}	R Documentation

Convert a data.tree structure to a list-of-list structure

Description

Convert a data.tree structure to a list-of-list structure

Usage

## S3 method for class 'Node'
as.list(x, mode = c("simple", "explicit"), unname = FALSE,
  nameName = ifelse(unname, "name", ""), childrenName = "children",
  rootName = "", keepOnly = NULL, ...)

ToListSimple(x, nameName = "name", ...)

ToListExplicit(x, unname = FALSE, nameName = ifelse(unname, "name", ""),
  childrenName = "children", ...)

Arguments

x
The Node to convert

mode
How the list is structured. "simple" (the default) will add children directly as nested lists. "explicit" puts children in a separate nested list called childrenName

unname
If TRUE, and if mode is "explicit", then the nested children list will not have named arguments. This can be useful e.g. in the context of conversion to JSON, if you prefer the children to be an array rather than named objects.

nameName
The name that should be given to the name element

childrenName
The name that should be given to the children nested list

rootName
The name of the node. If provided, this overrides Node$name

keepOnly
A character vector of fields to include in the result. If NULL (the default), all fields are kept.

...
Additional parameters passed to as.list.Node

Examples

data(acme)

str(ToListSimple(acme))
str(ToListSimple(acme, keepOnly = "cost"))

str(ToListExplicit(acme))
str(ToListExplicit(acme, unname = TRUE))
str(ToListExplicit(acme, unname = TRUE, nameName = "id", childrenName = "descendants"))

[Package data.tree version 0.7.4 Index]