as.phylo.Node {data.tree}	R Documentation

Convert a Node to a phylo object from the ape package.

Description

This method requires the ape package to be installed and loaded.

Usage

as.phylo.Node(x, heightAttribute = DefaultPlotHeight, ...)

Arguments

x
The root Node of the tree or sub-tree to be converted

heightAttribute
The attribute (field name or function) storing the height

...
any other argument

See Also

Other ape phylo conversions: GetPhyloNr, as.Node.phylo

Examples

library(ape)
data(acme)
acmephylo <- as.phylo(acme)
#plot(acmephylo)

[Package data.tree version 0.7.4 Index]